Moved Traits into seperate directory
