+++ src/Plugins/VulnerabilityScore.php
@@ -13,6 +13,7 @@ namespace Kanopi\Firewall\Plugins;
 
 use GeoIp2\Database\Reader;
 use GeoIp2\WebService\Client;
+use Kanopi\Firewall\Traits\GeoLocationTrait;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
