+++ src/Traits/EvaluateTrait.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Kanopi\Firewall\Plugins;
+namespace Kanopi\Firewall\Traits;
 
 use Kanopi\Firewall\Logging\LoggingTrait;
 use Symfony\Component\HttpFoundation\Request;
